Renaming files
@geekyme IMHO there's no need to change the case of the content we export, introducing a new naming rule.
